:art: consolidate files
